Redesign/fzn parser (#141)
* improved fzn redefinitions

* intermediate push

* monster commit

i am currently adding the new parser.
After the code could be compiled, many bugs occurred due to the interaction between the parser, invariant graph, invariant graph construction, and the unit tests.
I am thus currently redoing the invariant graph and invariant graph construction

* made InvariantGraph work with the new parser

* github actions cmake version

+++ test/constraints/tGlobalCardinalityConst.cpp
@@ -566,12 +566,12 @@ TEST_F(GlobalCardinalityTestClosed, EngineIntegration) {
       engine->open();
     }
     std::vector<VarId> args;
-    Int numArgs = 10;
-    for (Int value = 0; value < numArgs; ++value) {
+    size_t numArgs = 10;
+    for (size_t value = 0; value < numArgs; ++value) {
       args.push_back(engine->makeIntVar(0, -100, 100));
     }
 
-    VarId viol = engine->makeIntVar(0, 0, numArgs);
+    VarId viol = engine->makeIntVar(0, 0, static_cast<Int>(numArgs));
 
     testNotifications<MockGlobalCardinalityConst<false>>(
         &engine->makeInvariant<MockGlobalCardinalityConst<false>>(
@@ -586,12 +586,12 @@ TEST_F(GlobalCardinalityTestOpen, EngineIntegration) {
       engine->open();
     }
     std::vector<VarId> args;
-    Int numArgs = 10;
-    for (Int value = 0; value < numArgs; ++value) {
+    size_t numArgs = 10;
+    for (size_t value = 0; value < numArgs; ++value) {
       args.push_back(engine->makeIntVar(0, -100, 100));
     }
 
-    VarId viol = engine->makeIntVar(0, 0, numArgs);
+    VarId viol = engine->makeIntVar(0, 0, static_cast<Int>(numArgs));
 
     testNotifications<MockGlobalCardinalityConst<true>>(
         &engine->makeInvariant<MockGlobalCardinalityConst<true>>(
